Rename and split JUnit4Runner
