+++ infinitest-runner/src/main/java/org/infinitest/testrunner/SystemClock.java
@@ -30,7 +30,7 @@ package org.infinitest.testrunner;
 /**
  * The default system clock, which is used in production.
  */
-class SystemClock implements Clock {
+public class SystemClock implements Clock {
 	@Override
 	public long currentTimeMillis() {
 		return System.currentTimeMillis();
