+++ infinitest-runner/src/test/java/org/infinitest/testrunner/WhenRunningJUnit5Tests.java
@@ -1,0 +1,85 @@
+/*
+ * Infinitest, a Continuous Test Runner.
+ *
+ * Copyright (C) 2010-2013
+ * "Ben Rady" <benrady@gmail.com>,
+ * "Rod Coffin" <rfciii@gmail.com>,
+ * "Ryan Breidenbach" <ryan.breidenbach@gmail.com>
+ * "David Gageot" <david@gageot.net>, et al.
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+package org.infinitest.testrunner;
+
+import static com.google.common.collect.Iterables.getOnlyElement;
+import static org.infinitest.testrunner.TestEvent.methodFailed;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertTrue;
+
+import org.infinitest.testrunner.exampletests.FailingTest;
+import org.infinitest.testrunner.exampletests.JUnit5Test;
+import org.infinitest.testrunner.exampletests.PassingTestCase;
+import org.infinitest.testrunner.exampletests.TestNGTest;
+import org.infinitest.testrunner.junit5.Junit5Runner;
+import org.junit.After;
+import org.junit.Before;
+import org.junit.Test;
+
+import junit.framework.AssertionFailedError;
+
+public class WhenRunningJUnit5Tests {
+	private DefaultRunner runner;
+	
+	@Before
+	public void inContext() {
+		TestThatThrowsExceptionInConstructor.fail = true;
+		FailingTest.fail = true;
+		TestNGTest.fail = true;
+		runner = new DefaultRunner();
+	}
+
+	@After
+	public void cleanup() {
+		TestThatThrowsExceptionInConstructor.fail = false;
+		FailingTest.fail = false;
+	}
+
+	@Test
+	public void shouldSupportJUnit5() {
+		Iterable<TestEvent> events = runner.runTest(JUnit5Test.class.getName());
+		TestEvent expectedEvent = methodFailed(JUnit5Test.class.getName(), "shouldFail",
+				new AssertionFailedError("expected: <true> but was: <false>"));
+		assertEventsEquals(expectedEvent, getOnlyElement(events));
+	}
+
+	@Test
+	public void shouldDetectJUnit5Tests() {
+		assertTrue(Junit5Runner.isJUnit5Test(JUnit5Test.class));
+		assertFalse(Junit5Runner.isJUnit5Test(PassingTestCase.class));
+	}
+
+	private void assertEventsEquals(TestEvent expected, TestEvent actual) {
+		assertEquals(expected, actual);
+		assertEquals(expected.getMessage(), actual.getMessage());
+		assertEquals(expected.getType(), actual.getType());
+		assertEquals(expected.getErrorClassName(), actual.getErrorClassName());
+	}
+
+}
